src/1_input_processing/speech_to_text/requirements.txt
faster-whisper==1.0.0
pyannote-audio==2.1.0